servers/requirements.txt
mcp[cli]
anthropic
google-genai
exa-py
httpx
